[IMP] html_editor, *: TO SQUASH remove duplicates + move _can_bypass method to media
*: web_unsplash, website_forum
