+++ addons/html_editor/controllers/main.py
@@ -236,7 +236,11 @@ class HTML_Editor(http.Controller):
         context.pop('allowed_company_ids', None)
         request.update_env(context=context)
 
-    def _prepare_create(self, name='', data=False, url=False, res_id=False, res_model='ir.ui.view', media_type=None, **kwargs):
+    def _media_create(self, res_id, res_model, media_type=None, name='', data=False, url=False):
+        """Create and return a new media attachment."""
+        Media = request.env['html_editor.media']
+        IrAttachment = request.env['ir.attachment']
+
         if name.lower().endswith('.bmp'):
             # Avoid mismatch between content type and mimetype, see commit msg
             name = name[:-4]
@@ -282,20 +286,11 @@ class HTML_Editor(http.Controller):
         else:
             raise UserError(_("You need to specify either data or url to create an attachment."))
 
-        return attachment_data
-
-    def _media_create(self, res_id, res_model, media_type=None, name='', data=False, url=False):
-        """Create and return a new media attachment."""
-        Media = request.env['html_editor.media']
-        IrAttachment = request.env['ir.attachment']
-
-        attachment_data = self._prepare_create(name, data, url, res_id, res_model, media_type)
-
         # Despite the user having no right to create an attachment, he can still
         # create an image attachment through some flows
         if (
             not request.env.is_admin()
-            and IrAttachment._can_bypass_rights_on_media_dialog(**attachment_data)
+            and Media._can_bypass_rights_on_media_dialog(**attachment_data)
         ):
             media = Media.sudo().create(attachment_data)
             # When portal users upload an attachment with the wysiwyg widget,
@@ -318,31 +313,6 @@ class HTML_Editor(http.Controller):
             if not media:
                 media = Media.create(attachment_data)
         return media
-
-    # TODO @media-model: REMOVE
-    def _attachment_create(self, name='', data=False, url=False, res_id=False, res_model='ir.ui.view'):
-        """Create and return a new attachment."""
-        IrAttachment = request.env['ir.attachment']
-
-        attachment_data = self._prepare_create(name, data, url, res_id, res_model)
-        # Despite the user having no right to create an attachment, he can still
-        # create an image attachment through some flows
-        if (
-            not request.env.is_admin()
-            and IrAttachment._can_bypass_rights_on_media_dialog(**attachment_data)
-        ):
-            attachment = IrAttachment.sudo().create(attachment_data)
-            # When portal users upload an attachment with the wysiwyg widget,
-            # the access token is needed to use the image in the editor. If
-            # the attachment is not public, the user won't be able to generate
-            # the token, so we need to generate it using sudo
-            if not attachment_data['public']:
-                attachment.sudo().generate_access_token()
-        else:
-            attachment = get_existing_attachment(IrAttachment, attachment_data) \
-                or IrAttachment.create(attachment_data)
-
-        return attachment
 
     @http.route(['/web_editor/get_image_info', '/html_editor/get_image_info'], type='json', auth='user', website=True)
     def get_image_info(self, src=''):
@@ -388,21 +358,23 @@ class HTML_Editor(http.Controller):
             hide_dm_logo=hide_dm_logo, hide_dm_share=hide_dm_share
         )
 
-    def _preprocess_data(self, name, data, is_image, quality=0, width=0, height=0, **kwargs):
+    @http.route(['/web_editor/media/add_data', '/html_editor/media/add_data'], type='json', auth='user', methods=['POST'], website=True)
+    def add_media_data(self, name, data, is_image, quality=0, width=0, height=0, res_id=False, res_model='ir.ui.view', **kwargs):
+        """Creates an `html_editor.media` and its related `ir.attachment`."""
         data = b64decode(data)
-        preprocessed_data = {}
+        media_type = 'image' if is_image else 'document'
         if is_image:
             format_error_msg = _("Uploaded image's format is not supported. Try with: %s", ', '.join(SUPPORTED_IMAGE_MIMETYPES.values()))
             try:
                 data = tools.image_process(data, size=(width, height), quality=quality, verify_resolution=True)
-                preprocessed_data['mimetype'] = guess_mimetype(data)
-                if preprocessed_data['mimetype'] not in SUPPORTED_IMAGE_MIMETYPES:
+                mimetype = guess_mimetype(data)
+                if mimetype not in SUPPORTED_IMAGE_MIMETYPES:
                     return {'error': format_error_msg}
                 if not name:
                     name = '%s-%s%s' % (
                         datetime.now().strftime('%Y%m%d%H%M%S'),
                         str(uuid.uuid4())[:6],
-                        SUPPORTED_IMAGE_MIMETYPES[preprocessed_data['mimetype']],
+                        SUPPORTED_IMAGE_MIMETYPES[mimetype],
                     )
             except UserError:
                 # considered as an image by the browser file input, but not
@@ -411,40 +383,15 @@ class HTML_Editor(http.Controller):
             except ValueError as e:
                 return {'error': e.args[0]}
 
-        preprocessed_data.update({
-            'name': name,
-            'data': data,
-            'media_type': 'image' if is_image else 'document',
-        })
-        return preprocessed_data
-
-    @http.route(['/web_editor/media/add_data', '/html_editor/media/add_data'], type='json', auth='user', methods=['POST'], website=True)
-    def add_media_data(self, name, data, is_image, quality=0, width=0, height=0, res_id=False, res_model='ir.ui.view', **kwargs):
-        """Creates an `html_editor.media` and its related `ir.attachment`."""
-        preprocessed_data = self._preprocess_data(name, data, is_image, quality, width, height)
-        if 'error' in preprocessed_data:
-            return preprocessed_data
-
         self._clean_context()
         media = self._media_create(
             res_id,
             res_model,
-            media_type=preprocessed_data['media_type'],
-            name=preprocessed_data['name'],
-            data=preprocessed_data['data'],
+            media_type=media_type,
+            name=name,
+            data=data,
         )
         return media._get_media_info()
-
-    # TODO @media-model: REMOVE
-    @http.route(['/web_editor/attachment/add_data', '/html_editor/attachment/add_data'], type='json', auth='user', methods=['POST'], website=True)
-    def add_data(self, name, data, is_image, quality=0, width=0, height=0, res_id=False, res_model='ir.ui.view', **kwargs):
-        preprocessed_data = self._preprocess_data(name, data, is_image, quality, width, height)
-        if 'error' in preprocessed_data:
-            return preprocessed_data
-
-        self._clean_context()
-        attachment = self._attachment_create(name=preprocessed_data['name'], data=preprocessed_data['data'], res_id=res_id, res_model=res_model)
-        return attachment._get_media_info()
 
     @http.route(['/web_editor/media/add_url', '/html_editor/media/add_url'], type='json', auth='user', methods=['POST'], website=True)
     def add_url(self, url, res_id=False, res_model='ir.ui.view', **kwargs):
