+++ addons/html_editor/models/ir_attachment.py
@@ -71,19 +71,3 @@ class IrAttachment(models.Model):
             except UserError:
                 attachment.image_width = 0
                 attachment.image_height = 0
-
-    # TODO @media-model: adapt calls to this and remove this
-    def _get_media_info(self):
-        """Return a dict with the values that we need on the media dialog."""
-        self.ensure_one()
-        return self._read_format(['id', 'name', 'description', 'mimetype', 'checksum', 'url', 'type', 'res_id', 'res_model', 'public', 'access_token', 'image_src', 'image_width', 'image_height', 'original_id'])[0]
-
-    # TODO @media-model: move this to media model? Would make more sense
-    def _can_bypass_rights_on_media_dialog(self, **attachment_data):
-        """ This method is meant to be overridden, for instance to allow to
-        create image attachment despite the user not allowed to create
-        attachment, eg:
-        - Portal user uploading an image on the forum (bypass acl)
-        - Non admin user uploading an unsplash image (bypass binary/url check)
-        """
-        return False
