+++ addons/web_unsplash/models/__init__.py
@@ -3,5 +3,6 @@
 
 from . import ir_attachment
 from . import ir_qweb_fields
+from . import media
 from . import res_config_settings
 from . import res_users
