+++ addons/web_unsplash/models/ir_attachment.py
@@ -10,15 +10,6 @@ class Attachment(models.Model):
 
     _inherit = "ir.attachment"
 
-    def _can_bypass_rights_on_media_dialog(self, **attachment_data):
-        # We need to allow and sudo the case of an "url + file" attachment,
-        # which is by default forbidden for non admin.
-        # See `_check_serving_attachments`
-        forbidden = 'url' in attachment_data and attachment_data.get('type', 'binary') == 'binary'
-        if forbidden and attachment_data['url'].startswith('/unsplash/'):
-            return True
-        return super()._can_bypass_rights_on_media_dialog(**attachment_data)
-
     def _get_serve_attachment(self, url, extra_domain=None, order=None):
         # Bypass the default IrAttachment._search() behavior which adds
         # ('res_field', '=', False) to the domain if no `res_field` nor `id` is
